requirements.txt
streamlit
opencv-python
numpy
pillow
face-recognition